move hmac module from utils to top level
