+++ spylib/hmac.py
