+++ spylib/oauth/validations.py
@@ -3,8 +3,9 @@ from operator import itemgetter
 from typing import Any, List, Tuple
 from urllib.parse import parse_qsl
 
+from spylib.hmac import validate as validate_hmac
+
 from ..utils import domain_to_storename, now_epoch
-from ..utils import validate as validate_hmac
 from .tokens import OAuthJWT
