+++ spylib/utils/__init__.py
@@ -1,5 +1,4 @@
 from .domain import domain_to_storename, store_domain
-from .hmac import validate
 from .httpclient import HTTPClient
 from .jwtoken import JWTBaseModel
 from .misc import get_unique_id, now_epoch
@@ -8,7 +7,6 @@ from .rest import DELETE, GET, POST, PUT, Method
 __all__ = [
     'now_epoch',
     'get_unique_id',
-    'validate',
     'JWTBaseModel',
     'HTTPClient',
     'domain_to_storename',
